delete legacy files
