+++ .settings/org.eclipse.core.resources.prefs
@@ -1,4 +1,2 @@
 eclipse.preferences.version=1
-encoding//src/main/java=UTF-8
-encoding//src/test/java=UTF-8
 encoding/<project>=UTF-8
