+++ src/main/Jama/doc/resources/inherit.gif
