+++ src/main/Jama/examples/MagicSquareExample.java
