+++ src/main/Jama/util/Maths.java
