+++ src/main/LU.java
@@ -1,0 +1,16 @@
+
+public class LU {
+	
+	public static void main(String args[]) {
+		/* A is a random matrix of dimension dxd
+		 * L and U are the decomposed matrices of matrix A
+		 * C is the product of L and U (is identical to A)
+		 * B is a random vector of dimension d
+		 * x is the solution of the linear system Ax = B
+		 * Correctness verified
+		 * */
+		System.out.println("L:\n");
+		
+		
+	}
+}
